refactor Signed-off-by: Gabriel Fu <hfu.gabriel@gmail.com>
